apps/api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
redis==6.4.0
pydantic-settings==2.10.1
duckdb==1.4.0
psycopg2-binary==2.9.10
pandas==2.3.2
pyarrow==21.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
arq==0.26.1
